Refactor mod.rs files to clean module style
- Refactored parser/mod.rs to only contain module declarations and exports
- Refactored index/mod.rs to only contain module declarations and exports
- Moved implementation code to separate files:
  - parser/python_parser.rs: Contains PythonParser implementation
  - parser/extractor.rs: Contains SymbolExtractor implementation
  - index/symbol_index.rs: Contains SymbolIndex and FileMetadata implementations
- Fixed clippy warning about module inception by renaming parser.rs to python_parser.rs
- All tests passing

This follows the clean style of src/lsp/mod.rs where module files only declare and export other modules.
